control panel buttons added control panel renamed combined func comps added parameters to SemesterTable
